feat(T2): 回测可视化 HTML 报告，并暴露一个静默的仓位失真问题
## 可视化报告

新增 qmtquant/report/，产出单文件自包含 HTML（plotly.js 内联 4.1MB），
无需联网、无需服务，双击即可打开。

图表：净值曲线（可叠加基准）、回撤曲线（标注最大回撤区间）、
月度收益热力图（年×月，正红负绿）、单标的买卖点标注、
多标的的持仓只数变化与分标的盈亏排行。

报告同时呈现标的池 BiasReport —— 一个有幸存者偏差的漂亮曲线
比没有回测更危险，因为它会让人误以为验证过了。

已接入 run_backtest.py 与 run_portfolio_backtest.py。

## 发现的严重问题：后复权抬高价格导致标的完全买不进

后复权价远高于真实价（茅台 8152 vs 实际约 1300），一手的名义成本
被放大数倍，百股取整粒度因此变粗。

实测 100 万本金 / 10 只持仓（每只预算 9.5 万）：
  36 只（12%）取整后为 0 股，完全买不进 —— 000651、600519、300502 等
  另有 72 只闲置资金 >10%，最坏 49.8%

更糟的是它静默发生：策略原本在 rebalance() 里自行取整，不足一手就
continue 跳过，引擎不知情，报告也看不出这些标的被排除出实际持仓。

处理：
- 取整逻辑收敛到引擎一处，portfolio/trend_ma/ma_cross 不再自行取整
- 引擎统计 undersized_orders，报告以醒目告警呈现
  （实测组合回测触发 28 只标的受影响）
- lot_size 改为可配置

根本解法（未实现）：按真实价取整，需复权因子 F(t)=后复权价/真实价，
详见 docs/TASKS.md。

测试 196 passed（新增 22 个，覆盖图表数据变换与报告渲染，
含 XSS 转义与自包含性校验）

+++ qmtquant/strategy/examples/ma_cross.py
@@ -68,7 +68,7 @@ class MaCrossStrategy(StrategyBase):
         # 金叉且空仓 → 买入
         if prev_diff <= 0 < diff and pos == 0:
             cash = self.get_cash() * self.position_ratio
-            volume = int(cash / bar.close_price // 100) * 100
+            volume = cash / bar.close_price   # 整手约束由引擎处理
             if volume > 0:
                 # 用略高于收盘的限价，提高次日开盘成交概率
                 # 只在委托真正发出后才记日志：预热期 trading=False、
